+++ .env.local.example
@@ -1,0 +1,9 @@
+# Rename this file to .env.local and add your API keys
+
+# Claude API Key (Required for environmental tips generation)
+# Get your API key from: https://console.anthropic.com/account/keys
+REACT_APP_CLAUDE_API_KEY=your-claude-api-key-here
+
+# Supabase Configuration
+REACT_APP_SUPABASE_URL=your-supabase-url
+REACT_APP_SUPABASE_ANON_KEY=your-supabase-anon-key
