--- Ecostep 앱 완벽 재현 프롬프트.md
@@ -1,146 +1,0 @@
-# Ecostep 앱 완벽 재현 프롬프트
-
-React로 환경보호와 물고기 키우기가 결합된 모바일 앱 'Ecostep'을 만들어주세요.
-
-## 📱 전체 디자인 요구사항
-- 모바일 화면처럼 보이도록 핸드폰 프레임 구현 (390x844px, 검은색 베젤, 둥근 모서리 2.5rem)
-- 색상: 흰색, 검정색, 파란색(#3B82F6), 초록색(#10B981)만 사용
-- 다크모드 지원 (토글 가능)
-- 반응형 디자인 (가로는 max-w-[390px], 세로는 h-[844px] 고정)
-- 모든 텍스트는 작게 (text-xs, text-sm 위주)
-- 하단 네비게이션 아이콘은 선택시 채워지고, 미선택시 윤곽선만
-
-## 🏗️ 앱 구조
-5개 메인 탭: 홈, 챌린지, 보상, 커뮤니티, 기타
-상단바: "송일님 환영합니다" + 포인트(1240P) + 설정 아이콘
-
-## 📋 각 탭별 상세 기능
-
-### 1. 홈 탭
-- 어항 컴포넌트 (파란 배경, 물고기 이모지 🐠🐟🐡, 해초🌿, 산호🪸)
-- 어항 우측 상단에 설정 아이콘 (클릭시 어항 설정 화면)
-- 수질 표시 바 (85%)
-- "🔥 23일 연속 달성!" 알림 (초록색 배경)
-- 플라스틱 절약량 카드: 18.7kg 표시
-- min-height로 화면 꽉 채우기
-
-### 2. 챌린지 탭
-서브탭: [습관 챌린지] [플라스틱 추적]
-
-#### 습관 챌린지
-- 챌린지 선택 드롭다운 (기본: "플라스틱 빨대 안쓰기")
-- "기타 (직접 입력)" 선택시 커스텀 챌린지 추가 가능
-- 7일 진행 표시 (1-3일: 초록 체크, 4일: 빨간 X, 5-7일: 회색 원)
-- 진행률 바 57%
-- "오늘 완료하기 (+10 포인트)" 버튼
-- 지난 챌린지: 기본 3개만 표시, "더보기" 버튼으로 전체 표시
-- 각 챌린지마다 진행률 바 표시
-
-#### 플라스틱 추적
-- 이번 주 목표 설정 (500g)
-- 플라스틱 사용 기록하기:
-  - 아이템 선택 (플라스틱병 20g, 음식용기 30g, 컵 15g, 비닐봉지 5g, 기타)
-  - "기타" 선택시: 항목 이름, 개당 무게 입력 가능
-  - 수량 입력시 자동 무게 계산
-- 사용량 분석 (막대그래프)
-- 주간 사용량 추이 (세로 막대 차트)
-- 이번주 기록 (월~일 전체 표시)
-
-### 3. 보상 탭
-- 현재 랭크 카드: Gold 표시, 플래티넘까지 70% 진행률
-- 랭크 아이콘: 🥉브론즈 🥈실버 🥇골드 👑플래티넘
-- 랭킹 보상 (3개 카드 나란히)
-- 구분선
-- "물고기" 섹션:
-  - 브론즈: 금붕어, 구피, 네온테트라 (각 100-300포인트)
-  - 실버: 엔젤피시, 디스커스, 베타 (각 300-500포인트)
-  - 골드: 아로와나, 플라워혼, 블루탱 (각 500-700포인트)
-  - 플래티넘: 아시안아로와나, 플래티넘아로와나, 폴카닷스팅레이 (잠금/블러)
-- 구매 완료된 물고기: 초록색 테두리, "구매완료" 표시
-- 구분선
-- 어항 장식품: 해초 50P, 산호 80P, 성 100P
-
-### 4. 커뮤니티 탭
-- 친구 초대: [카톡으로 초대] [링크 복사] 버튼
-- 친구 검색: 돋보기 아이콘 + "아이디로 친구 검색" 입력란
-- 내 친구 랭킹 TOP 3 (나는 파란 링으로 강조)
-- 전체 랭킹 (나는 "상위 12%" 표시)
-
-### 5. 기타 탭
-- 환경 뉴스 (3개 항목, 날짜와 "더보기 →")
-- 제로웨이스트 맵 (지도 아이콘, 근처 장소 3개)
-- 도움말 섹션:
-  - 📚 앱 사용 방법
-  - 📞 고객 센터
-
-## ⚙️ 설정 화면
-메인 설정에서 모든 항목이 동일한 크기의 카드
-- 송일님의 정보 (프로필 카드)
-- 언어 (한국어/English)
-- 알림 (온/오프 토글)
-- 화면 테마 (라이트/다크)
-
-### 프로필 화면
-- 프로필 사진 업로드 영역 (카메라 아이콘 + 버튼)
-- 이름: 송일
-- 아이디: @songil_eco (친구 찾기용)
-- 생년월일: 2004.12.02
-- 이메일: callmesongil@kakao.com
-- 휴대폰: 010-5633-3473
-- 계정 탈퇴 (빨간색)
-
-## 🐠 어항 설정 화면
-- 어항 선택 (실버/골드/플래티넘)
-- 물고기 수 설정 (1-12마리, +/- 버튼)
-- 랜덤 체크박스 (체크시 자동 선택, 미체크시 수동 선택)
-- 랭크별로 구매한 물고기만 표시
-- 어항 꾸미기: 해초, 산호, 성, 돌, 조개, 해마상
-
-## 💾 상태 관리
-useState 사용:
-- 포인트: 1240
-- 다크모드 토글
-- 구매한 물고기 배열: ['금붕어', '구피', '네온테트라', '엔젤피시']
-- 커스텀 챌린지/플라스틱 항목 배열
-- 챌린지 진행 상태
-
-## 🎨 특별 요구사항
-1. 플래티넘 물고기는 블러 처리, 이름은 한 줄로 표시 (text-[10px])
-2. 모든 설정 화면은 별도 컴포넌트로 구성
-3. 정보가 적을 때도 min-h-full로 화면 채우기
-4. 지난 챌린지는 max-h-60 overflow-y-auto로 스크롤
-5. 반드시 lucide-react 아이콘 사용
-6. 토스 앱의 다크모드 디자인 참고하여 깔끔하고 모던한 느낌
-
-## 🔧 기술 스택
-- React (함수형 컴포넌트)
-- Tailwind CSS
-- lucide-react (아이콘)
-- useState Hook
-
-## 📝 추가 기능 상세
-
-### 챌린지 기능
-- 커스텀 챌린지 추가 시 배열에 저장
-- 진행 상태는 일별로 체크/X/미완료 표시
-- 지난 챌린지는 완료율 퍼센트 표시
-
-### 플라스틱 추적
-- 기타 항목 선택시 이름과 무게 커스터마이징
-- 수량 × 개당무게 = 총무게 자동계산
-- 주간 데이터 시각화
-
-### 물고기 시스템
-- 랭크별 3마리씩 총 12마리
-- 구매한 물고기만 어항 설정에서 선택 가능
-- 구매 완료 상태 저장 및 표시
-
-### 다크모드
-- 모든 컴포넌트에 조건부 스타일링
-- bgColor, textColor, borderColor, cardBg, inputBg 변수 사용
-
-## 📌 중요 포인트
-- 모바일 퍼스트 디자인
-- 직관적인 UX/UI
-- 게이미피케이션 요소 (포인트, 랭크, 물고기)
-- 환경보호 동기부여 (통계, 진행률, 연속 달성)
